Removed the meta directories: 	"460Demo.xcodeproj/project.xcworkspace/xcshareddata/" 	"460Demo.xcodeproj/project.xcworkspace/xcuserdata/" 	"460Demo.xcodeproj/xcuserdata/"
These directories are annoying because we didn't change any files in our project but since Xcode put changes in these directory, there will be uncommitted changes and we either have to commit this file or discard the change everytime before we do a git pull.

Also added a .gitignore to ignore the above directories.